+++ sotorrent/posttags/bigquery/PostTags.sql
@@ -1,6 +1,6 @@
 SELECT temp.PostId AS PostId, tags.Id AS TagId
-FROM `sotorrent-org.2020_08_31.Tags` tags
-JOIN `sotorrent-org.2020_08_31.PostTagsTemp` temp
+FROM `sotorrent-org.2020_11_12.Tags` tags
+JOIN `sotorrent-org.2020_11_12.PostTagsTemp` temp
 ON tags.TagName = temp.Tag;
 
-=> `sotorrent-org.2020_06_31.PostTags`
+=> `sotorrent-org.2020_11_12.PostTags`