+++ sotorrent/upload/biguery/import-tables.sh
@@ -1,12 +1,15 @@
 #!/bin/bash
 
 project="sotorrent-org"
-dataset="2020_08_31"
+dataset="2020_11_16"
 bucket="sotorrent"
 logfile="bigquery.log"
 gcloud_mysql_instance="sotorrent"
-gcloud_mysql_db="sotorrent20_06"
-glcoud_mysql_connection="sotorrent-org.US.sotorrent20_06"
+gcloud_mysql_db="sotorrent20_09"
+glcoud_mysql_connection="sotorrent-org.US.sotorrent20_09"
+
+# test database connection
+#gcloud sql connect $gcloud_mysql_instance --user=sotorrent;
 
 load_so=false
 load_gh=false
@@ -39,112 +42,95 @@ if [ "$2" = "delete-csv" ]; then
 fi
 
 if [ "$load_so" = true ] ; then 
-	echo "Decompressing $1 files..." | tee -a "$log_file"
-	for file in ./so-dump-bigquery/*.7z; do 7z e "$file" -o./so-dump/; done
+	echo "Decompressing $1 files..." | tee -a "$logfile"
+	for file in ./so-dump/*.7z; do 7z e "$file" -o./so-dump/; done
 	
-	# upload SQL dumps to Google Cloud Bucket
-	echo "Uploading $1 SQL dumps..." | tee -a "$log_file"
-	gsutil cp so-dump-bigquery/Users.sql gs://$bucket/
-	gsutil cp so-dump-bigquery/Badges.sql gs://$bucket/
-	gsutil cp so-dump-bigquery/PostLinks.sql gs://$bucket/
-	gsutil cp so-dump-bigquery/Tags.sql gs://$bucket/
-	gsutil cp so-dump-bigquery/Votes.sql gs://$bucket/
-	gsutil cp so-dump-bigquery/Comments.sql gs://$bucket/
-	gsutil cp so-dump-bigquery/Posts.sql gs://$bucket/
-	gsutil cp so-dump-bigquery/PostHistory.sql gs://$bucket/
-
-	# test database connection
-	#gcloud sql connect $gcloud_mysql_instance --user=sotorrent;
-
-	# import SQL dumps into Google Cloud MySQL instance
-
-	echo "Importing $1 SQL dumps into MySQL instance..." | tee -a "$log_file"
+	echo "Uploading $1 SQL dumps, importing them into MySQL instance, and copying them into BigQuery dataset..." | tee -a "$logfile"
 	
-	gcloud sql import sql --quiet \
-	  $gcloud_mysql_instance \
-	  gs://$bucket/Badges.sql \
-	  --database=$gcloud_mysql_db;
-	
-	sleep 1h
-	
-	gcloud sql import sql --quiet \
-	  $gcloud_mysql_instance \
-	  gs://$bucket/PostLinks.sql \
-	  --database=$gcloud_mysql_db;
-	
-	sleep 1h
-	
-	gcloud sql import sql --quiet \
-	  $gcloud_mysql_instance \
-	  gs://$bucket/Tags.sql \
-	  --database=$gcloud_mysql_db;
-	
-	sleep 1h
+	gsutil cp so-dump/Users.sql gs://$bucket/	
 	
 	gcloud beta sql import sql --quiet \
 	  $gcloud_mysql_instance \
 	  gs://$bucket/Users.sql \
 	  --database=$gcloud_mysql_db;
 	
-	sleep 1h
+	gsutil cp so-dump/Badges.sql gs://$bucket/
+	
+	# at this point the import into MySQL is hopefully finished
+	bq query --nouse_legacy_sql \
+	"CREATE TABLE \`$dataset.Users\`
+	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM Users;\");"
+	
+	gcloud sql import sql --quiet \
+	  $gcloud_mysql_instance \
+	  gs://$bucket/Badges.sql \
+	  --database=$gcloud_mysql_db;
+	
+	gsutil cp so-dump/PostLinks.sql gs://$bucket/
+	
+	bq query --nouse_legacy_sql \
+	"CREATE TABLE \`$dataset.Badges\`
+	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM Badges;\");"
+	
+	gcloud sql import sql --quiet \
+	  $gcloud_mysql_instance \
+	  gs://$bucket/PostLinks.sql \
+	  --database=$gcloud_mysql_db;
+	
+	gsutil cp so-dump/Tags.sql gs://$bucket/
+	
+	bq query --nouse_legacy_sql \
+	"CREATE TABLE \`$dataset.PostLinks\`
+	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM PostLinks;\");"
+	
+	gcloud sql import sql --quiet \
+	  $gcloud_mysql_instance \
+	  gs://$bucket/Tags.sql \
+	  --database=$gcloud_mysql_db;
+	
+	gsutil cp so-dump/Votes.sql gs://$bucket/
+	
+	bq query --nouse_legacy_sql \
+	"CREATE TABLE \`$dataset.Tags\`
+	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM Tags;\");"
 	
 	gcloud sql import sql --quiet \
 	  $gcloud_mysql_instance \
 	  gs://$bucket/Votes.sql \
 	  --database=$gcloud_mysql_db;
 	
-	sleep 1h
+	gsutil cp so-dump/Comments.sql gs://$bucket/
+	
+	bq query --nouse_legacy_sql \
+	"CREATE TABLE \`$dataset.Votes\`
+	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM Votes;\");"
 	
 	gcloud sql import sql --quiet \
 	  $gcloud_mysql_instance \
 	  gs://$bucket/Comments.sql \
 	  --database=$gcloud_mysql_db;
 	
-	sleep 1h
+	gsutil cp so-dump/Posts.sql gs://$bucket/
+	
+	bq query --nouse_legacy_sql \
+	"CREATE TABLE \`$dataset.Comments\`
+	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM Comments;\");"
 	
 	gcloud sql import sql --quiet \
 	  $gcloud_mysql_instance \
 	  gs://$bucket/Posts.sql \
 	  --database=$gcloud_mysql_db;
 	
-	sleep 1.5h
+	gsutil cp so-dump/PostHistory.sql gs://$bucket/
+	
+	bq query --nouse_legacy_sql \
+	"CREATE TABLE \`$dataset.Posts\`
+	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM Posts;\");"
 	
 	gcloud sql import sql --quiet \
 	  $gcloud_mysql_instance \
 	  gs://$bucket/PostHistory.sql \
 	  --database=$gcloud_mysql_db;
-	
-	sleep 2h
-	
-	echo "Copying $1 tables from MySQL instance to BigQuery..." | tee -a "$log_file"
-	
-		bq query --nouse_legacy_sql \
-	"CREATE TABLE \`$dataset.Badges\`
-	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM Badges;\");"
-	
-		bq query --nouse_legacy_sql \
-	"CREATE TABLE \`$dataset.PostLinks\`
-	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM PostLinks;\");"
-	
-		bq query --nouse_legacy_sql \
-	"CREATE TABLE \`$dataset.Tags\`
-	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM Tags;\");"
-	
-		bq query --nouse_legacy_sql \
-	"CREATE TABLE \`$dataset.Users\`
-	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM Users;\");"
-	
-		bq query --nouse_legacy_sql \
-	"CREATE TABLE \`$dataset.Votes\`
-	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM Votes;\");"
-	
-	bq query --nouse_legacy_sql \
-	"CREATE TABLE \`$dataset.Comments\`
-	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM Comments;\");"
-	
-	bq query --nouse_legacy_sql \
-	"CREATE TABLE \`$dataset.Posts\`
-	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM Posts;\");"
 	
 	bq query --nouse_legacy_sql \
 	"CREATE TABLE \`$dataset.PostHistory\`
@@ -159,172 +145,147 @@ if [ "$load_so" = true ] ; then
 fi
 
 if [ "$load_sotorrent" = true ] ; then 
-	echo "Decompressing $1 files..." | tee -a "$log_file"
+	echo "Decompressing $1 files..." | tee -a "$logfile"
 	for file in ./sotorrent/*.7z; do 7z e "$file" -o./sotorrent/; done
-
-	echo "Uploading $1 SQL dumps..." | tee -a "$log_file"
-	gsutil cp sotorrent/PostBlockDiff.sql gs://$bucket/
+	
+	echo "Uploading $1 SQL dumps, importing them into MySQL instance, and copying them into BigQuery dataset..." | tee -a "$logfile"
+	
 	gsutil cp sotorrent/PostVersion.sql gs://$bucket/
-	gsutil cp sotorrent/PostBlockVersion.sql gs://$bucket/
-	gsutil cp sotorrent/PostVersionUrl.sql gs://$bucket/
-	gsutil cp sotorrent/CommentUrl.sql gs://$bucket/
-	gsutil cp sotorrent/TitleVersion.sql gs://$bucket/ 
-	gsutil cp sotorrent/StackSnippetVersion.sql gs://$bucket/   
-	gsutil cp sotorrent/PostViews.sql gs://$bucket/   
-	gsutil cp sotorrent/PostTags.sql gs://$bucket/   
-	
-	echo "Importing $1 SQL dumps into MySQL instance..." | tee -a "$log_file"
-	
-	gcloud sql import sql --quiet \
-	  $gcloud_mysql_instance \
-	  gs://$bucket/PostBlockDiff.sql \
-	  --database=$gcloud_mysql_db;
-	
-	sleep 1.5h
 	
 	gcloud sql import sql --quiet \
 	  $gcloud_mysql_instance \
 	  gs://$bucket/PostVersion.sql \
 	  --database=$gcloud_mysql_db;
 	
-	sleep 1h
+	gsutil cp sotorrent/PostVersionUrl.sql gs://$bucket/
 	
-	gcloud sql import sql --quiet \
-	  $gcloud_mysql_instance \
-	  gs://$bucket/PostBlockVersion.sql \
-	  --database=$gcloud_mysql_db;
-	
-	sleep 5h
+	bq query --nouse_legacy_sql \
+	"CREATE TABLE \`$dataset.PostVersion\`
+	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM PostVersion;\");"
 	
 	gcloud sql import sql --quiet \
 	  $gcloud_mysql_instance \
 	  gs://$bucket/PostVersionUrl.sql \
 	  --database=$gcloud_mysql_db;
 	
-	sleep 1h
+	gsutil cp sotorrent/CommentUrl.sql gs://$bucket/
+	
+	bq query --nouse_legacy_sql \
+	"CREATE TABLE \`$dataset.PostVersionUrl\`
+	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM PostVersionUrl;\");"
 	
 	gcloud sql import sql --quiet \
 	  $gcloud_mysql_instance \
 	  gs://$bucket/CommentUrl.sql \
 	  --database=$gcloud_mysql_db;
 	
-	sleep 1h
+	gsutil cp sotorrent/TitleVersion.sql gs://$bucket/
+	
+	bq query --nouse_legacy_sql \
+	"CREATE TABLE \`$dataset.CommentUrl\`
+	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM CommentUrl;\");"
 	
 	gcloud sql import sql --quiet \
 	  $gcloud_mysql_instance \
 	  gs://$bucket/TitleVersion.sql \
 	  --database=$gcloud_mysql_db;
 	
-	sleep 1h
+	gsutil cp sotorrent/StackSnippetVersion.sql gs://$bucket/   
 	
-	gcloud sql import sql --quiet \
-	  $gcloud_mysql_instance \
-	  gs://$bucket/TitleVersion.sql \
-	  --database=$gcloud_mysql_db;
-	
-	sleep 1h
+	bq query --nouse_legacy_sql \
+	"CREATE TABLE \`$dataset.TitleVersion\`
+	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM TitleVersion;\");"
 	
 	gcloud sql import sql --quiet \
 	  $gcloud_mysql_instance \
 	  gs://$bucket/StackSnippetVersion.sql \
 	  --database=$gcloud_mysql_db;
 	
-	sleep 1h
+	gsutil cp sotorrent/PostViews.sql gs://$bucket/   
+	
+	bq query --nouse_legacy_sql \
+	"CREATE TABLE \`$dataset.StackSnippetVersion\`
+	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM StackSnippetVersion;\");"
 	
 	gcloud sql import sql --quiet \
 	  $gcloud_mysql_instance \
 	  gs://$bucket/PostViews.sql \
 	  --database=$gcloud_mysql_db;
 	
-	sleep 1h
+	gsutil cp sotorrent/PostTags.sql gs://$bucket/  
+	
+	bq query --nouse_legacy_sql \
+	"CREATE TABLE \`$dataset.PostViews\`
+	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM PostViews;\");"
 	
 	gcloud sql import sql --quiet \
 	  $gcloud_mysql_instance \
 	  gs://$bucket/PostTags.sql \
 	  --database=$gcloud_mysql_db;
 	
-	sleep 1h
+	gsutil cp sotorrent/PostBlockDiff.sql gs://$bucket/
 	
-	echo "Copying $1 tables from MySQL instance to BigQuery..." | tee -a "$log_file"
+	bq query --nouse_legacy_sql \
+	"CREATE TABLE \`$dataset.PostTags\`
+	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM PostTags;\");"
+	
+	gcloud sql import sql --quiet \
+	  $gcloud_mysql_instance \
+	  gs://$bucket/PostBlockDiff.sql \
+	  --database=$gcloud_mysql_db;
+	
+	gsutil cp sotorrent/PostBlockVersion.sql gs://$bucket/
 	
 	bq query --nouse_legacy_sql \
 	"CREATE TABLE \`$dataset.PostBlockDiff\`
 	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM PostBlockDiff;\");"
 	
-	bq query --nouse_legacy_sql \
-	"CREATE TABLE \`$dataset.PostVersion\`
-	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM PostVersion;\");"
+	gcloud sql import sql --quiet \
+	  $gcloud_mysql_instance \
+	  gs://$bucket/PostBlockVersion.sql \
+	  --database=$gcloud_mysql_db;
+	
+	sleep 3h
 	
 	bq query --nouse_legacy_sql \
 	"CREATE TABLE \`$dataset.PostBlockVersion\`
 	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM PostBlockVersion;\");"
-	
-	bq query --nouse_legacy_sql \
-	"CREATE TABLE \`$dataset.PostVersionUrl\`
-	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM PostVersionUrl;\");"
-	
-	bq query --nouse_legacy_sql \
-	"CREATE TABLE \`$dataset.CommentUrl\`
-	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM CommentUrl;\");"
-	
-	bq query --nouse_legacy_sql \
-	"CREATE TABLE \`$dataset.TitleVersion\`
-	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM TitleVersion;\");"
-	
-	bq query --nouse_legacy_sql \
-	"CREATE TABLE \`$dataset.StackSnippetVersion\`
-	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM StackSnippetVersion;\");"
-	
-	bq query --nouse_legacy_sql \
-	"CREATE TABLE \`$dataset.PostViews\`
-	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM PostViews;\");"
-	
-	bq query --nouse_legacy_sql \
-	"CREATE TABLE \`$dataset.PostTags\`
-	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM PostTags;\");"
 fi
 
 if [ "$load_gh" = true ] ; then 
-	echo "Decompressing $1 SQL dumps..." | tee -a "$log_file"
+	echo "Decompressing $1 SQL dumps..." | tee -a "$logfile"
 	for file in ./gh-references/*.7z; do 7z e "$file" -o./gh-references/; done
 	
-	echo "Uploading $1 tables..." | tee -a "$log_file"
+	echo "Uploading $1 SQL dumps, importing them into MySQL instance, and copying them into BigQuery dataset..." | tee -a "$logfile"
+	
 	gsutil cp gh-references/PostReferenceGH.sql gs://$bucket/
-	gsutil cp gh-references/GHMatches.sql gs://$bucket/
-	gsutil cp gh-references/GHCommits.sql gs://$bucket/
-
-	echo "Importing $1 SQL dumps into MySQL instance..." | tee -a "$log_file"
 	
 	gcloud sql import sql --quiet \
 	  $gcloud_mysql_instance \
 	  gs://$bucket/PostReferenceGH.sql \
 	  --database=$gcloud_mysql_db;
 	
-	sleep 1h
+	gsutil cp gh-references/GHMatches.sql gs://$bucket/
+	
+	bq query --nouse_legacy_sql \
+	"CREATE TABLE \`$dataset.PostReferenceGH\`
+	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM PostReferenceGH;\");"
 	
 	gcloud sql import sql --quiet \
 	  $gcloud_mysql_instance \
 	  gs://$bucket/GHMatches.sql \
 	  --database=$gcloud_mysql_db;
 	
-	sleep 1h
-	
-		gcloud sql import sql --quiet \
-	  $gcloud_mysql_instance \
-	  gs://$bucket/GHCommits.sql \
-	  --database=$gcloud_mysql_db;
-	
-	sleep 1h
-	
-	echo "Copying $1 tables from MySQL instance to BigQuery..." | tee -a "$log_file"
-	
-	bq query --nouse_legacy_sql \
-	"CREATE TABLE \`$dataset.PostReferenceGH\`
-	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM PostReferenceGH;\");"
+	gsutil cp gh-references/GHCommits.sql gs://$bucket/
 	
 	bq query --nouse_legacy_sql \
 	"CREATE TABLE \`$dataset.GHMatches\`
 	AS SELECT * FROM EXTERNAL_QUERY(\"$glcoud_mysql_connection\", \"SELECT * FROM GHMatches;\");"
+	
+	gcloud sql import sql --quiet \
+	  $gcloud_mysql_instance \
+	  gs://$bucket/GHCommits.sql \
+	  --database=$gcloud_mysql_db;
 	
 	bq query --nouse_legacy_sql \
 	"CREATE TABLE \`$dataset.GHCommits\`