+++ sotorrent/upload/zenodo/3_calculate-md5.sh
@@ -2,23 +2,23 @@
 
 # on some systems, you need to replace md5sum by md5
 
-md5sum so-dump/Badges.xml.7z
-md5sum so-dump/Comments.xml.7z
+md5sum so-dump/Badges.sql.7z
+md5sum so-dump/Comments.sql.7z
 md5sum sotorrent/CommentUrl.sql.7z
 md5sum gh-references/GHCommits.sql.7z
 md5sum gh-references/GHMatches.sql.7z
 md5sum sotorrent/PostBlockDiff.sql.7z
 md5sum sotorrent/PostBlockVersion.sql.7z
-md5sum so-dump/PostHistory.xml.7z
-md5sum so-dump/PostLinks.xml.7z
+md5sum so-dump/PostHistory.sql.7z
+md5sum so-dump/PostLinks.sql.7z
 md5sum gh-references/PostReferenceGH.sql.7z
-md5sum so-dump/Posts.xml.7z
+md5sum so-dump/Posts.sql.7z
 md5sum sotorrent/PostTags.sql.7z
 md5sum sotorrent/PostVersion.sql.7z
 md5sum sotorrent/PostVersionUrl.sql.7z
 md5sum sotorrent/PostViews.sql.7z
 md5sum sotorrent/StackSnippetVersion.sql.7z
-md5sum so-dump/Tags.xml.7z
+md5sum so-dump/Tags.sql.7z
 md5sum sotorrent/TitleVersion.sql.7z
-md5sum so-dump/Users.xml.7z
-md5sum so-dump/Votes.xml.7z
+md5sum so-dump/Users.sql.7z
+md5sum so-dump/Votes.sql.7z